Move main_screen.dart to /screens
