+++ app/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:locksense/services/firebase_service.dart';
-import './main_screen.dart';
+import '../screens/main_screen.dart';
 import './theme/style.dart';
 import './firebase_options.dart';
